Added Book, pdf, document, and image size to matrix page.  Updated Fields to be Long.
git-tfs-id: [https://tfsnpt.int.thomson.com:443/tfs/Cobalt_Collection/]$/Cobalt eBook Builder/Development;C222564

+++ ebookGenerator/src/com/thomsonreuters/uscl/ereader/assemble/step/AssembleEbook.java
@@ -84,10 +84,10 @@ public class AssembleEbook extends AbstractSbTasklet {
 		
 		PublishingStats jobstatsFormat = new PublishingStats();
 		jobstatsFormat.setJobInstanceId(jobId);
-		jobstatsFormat.setLargestDocSize((int)largestDocuent);
-		jobstatsFormat.setLargestImageSize((int)largestImage);
-		jobstatsFormat.setLargestPdfSize((int)largestPdf);
-		jobstatsFormat.setBookSize((int)gzipeSize);
+		jobstatsFormat.setLargestDocSize((long)largestDocuent);
+		jobstatsFormat.setLargestImageSize((long)largestImage);
+		jobstatsFormat.setLargestPdfSize((long)largestPdf);
+		jobstatsFormat.setBookSize((long)gzipeSize);
 		
 		
 		LOG.debug("largestDocuent ="+largestDocuent);
